+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Extra options for Text - `hide_on_load`, `dom_id` and `css_classes`.
+- Text control can display XML with syntax highlighting.
 ### Changed
 ### Fixed